+++ src/lib/components/table/codes/CodesTableMobileCard.tsx
@@ -8,13 +8,13 @@ import { PermissionChip } from "lib/components/PermissionChip";
 import type { CodeInfo } from "lib/types";
 import { getCw2Info } from "lib/utils";
 
-import { CodeNameCell } from "./CodeNameCell";
-
 interface CodesTableMobileCardProps {
   codeInfo: CodeInfo;
+  showCw2andContracts: boolean;
 }
 export const CodesTableMobileCard = ({
   codeInfo,
+  showCw2andContracts,
 }: CodesTableMobileCardProps) => {
   const cw2Info = getCw2Info(codeInfo.cw2Contract, codeInfo.cw2Version);
   const navigate = useInternalNavigate();
@@ -28,21 +28,38 @@ export const CodesTableMobileCard = ({
         })
       }
       topContent={
-        <Flex gap={2} align="center">
-          <MobileLabel variant="body2" label="Code ID" />
-          <ExplorerLink
-            type="code_id"
-            value={codeInfo.id.toString()}
-            showCopyOnHover
-          />
-        </Flex>
+        showCw2andContracts ? (
+          <Flex gap={2} align="center">
+            <MobileLabel variant="body2" label="Code ID" />
+            <ExplorerLink
+              type="code_id"
+              value={codeInfo.id.toString()}
+              showCopyOnHover
+            />
+          </Flex>
+        ) : (
+          <Flex gap={3} w="full">
+            <Flex direction="column" flex="1">
+              <MobileLabel variant="body2" label="Code ID" />
+              <ExplorerLink
+                type="code_id"
+                value={codeInfo.id.toString()}
+                showCopyOnHover
+              />
+            </Flex>
+            <Flex direction="column" flex="1" gap={1}>
+              <MobileLabel variant="body2" label="Permission" />
+              <PermissionChip
+                instantiatePermission={codeInfo.instantiatePermission}
+                permissionAddresses={codeInfo.permissionAddresses}
+                tagSize="xs"
+              />
+            </Flex>
+          </Flex>
+        )
       }
       middleContent={
-        <Flex direction="column" gap={3}>
-          <Flex direction="column">
-            <MobileLabel label="Code Name" />
-            <CodeNameCell code={codeInfo} isReadOnly />
-          </Flex>
+        showCw2andContracts && (
           <Flex direction="column">
             <MobileLabel label="CW2 Info" />
             <Text
@@ -53,30 +70,32 @@ export const CodesTableMobileCard = ({
               {cw2Info ?? "N/A"}
             </Text>
           </Flex>
-        </Flex>
+        )
       }
       bottomContent={
-        <Flex gap={3} w="full">
-          <Flex direction="column" flex="1">
-            <MobileLabel label="Contracts" />
-            <Text
-              variant="body3"
-              onClick={(e) => e.stopPropagation()}
-              cursor="text"
-              color={codeInfo.contractCount ? "text.main" : "text.disabled"}
-            >
-              {codeInfo.contractCount ?? "N/A"}
-            </Text>
+        showCw2andContracts && (
+          <Flex gap={3} w="full">
+            <Flex direction="column" flex="1">
+              <MobileLabel label="Contracts" />
+              <Text
+                variant="body3"
+                onClick={(e) => e.stopPropagation()}
+                cursor="text"
+                color={codeInfo.contractCount ? "text.main" : "text.disabled"}
+              >
+                {codeInfo.contractCount ?? "N/A"}
+              </Text>
+            </Flex>
+            <Flex direction="column" flex="1">
+              <MobileLabel label="Permission" />
+              <PermissionChip
+                instantiatePermission={codeInfo.instantiatePermission}
+                permissionAddresses={codeInfo.permissionAddresses}
+                tagSize="xs"
+              />
+            </Flex>
           </Flex>
-          <Flex direction="column" flex="1">
-            <MobileLabel label="Permission" />
-            <PermissionChip
-              instantiatePermission={codeInfo.instantiatePermission}
-              permissionAddresses={codeInfo.permissionAddresses}
-              tagSize="xs"
-            />
-          </Flex>
-        </Flex>
+        )
       }
     />
   );