+++ src/lib/pages/code-details/index.tsx
@@ -16,7 +16,7 @@ import PageContainer from "lib/components/PageContainer";
 import { ErrorFetching, InvalidState } from "lib/components/state";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 import { useSchemaStore } from "lib/providers/store";
-import { useCodeDataByCodeId } from "lib/services/codeService";
+import { useCodeDataByCodeId } from "lib/services/wasm/code";
 
 import { CodeContractsTable, CodeInfoSection } from "./components/code-info";
 import { CodeTopInfo } from "./components/code-info/CodeTopInfo";