+++ src/lib/pages/migrate/index.tsx
@@ -18,7 +18,7 @@ import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
 import { useUploadCode } from "lib/hooks";
 import { useContractDetailByContractAddress } from "lib/services/contractService";
-import { useUploadAccessParams } from "lib/services/proposalService";
+import { useUploadAccessParams } from "lib/services/wasm/proposal";
 import type { BechAddr32 } from "lib/types";
 import { getFirstQueryParam } from "lib/utils";