+++ src/lib/services/searchService.ts
@@ -20,14 +20,14 @@ import {
 } from "lib/utils";
 
 import { useBlockData } from "./blockService";
-import { useCodeDataByCodeId } from "./codeService";
 import { queryContract } from "./contract";
 import { useModuleByAddressLcd } from "./move/moduleService";
 import { useAddressByICNSName, useICNSNamesByAddress } from "./nameService";
 import type { ICNSNamesResponse } from "./ns";
 import { usePoolByPoolId } from "./poolService";
-import { useProposalData } from "./proposalService";
 import { useValidatorData } from "./validatorService";
+import { useCodeDataByCodeId } from "./wasm/code";
+import { useProposalData } from "./wasm/proposal";
 import { useTxData } from "./wasm/txs";
 
 export type SearchResultType =