+++ src/lib/services/wasm/code/api.ts
@@ -1,0 +1,105 @@
+import type { UseQueryOptions, UseQueryResult } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
+import axios from "axios";
+
+import {
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useGovConfig,
+} from "lib/app-provider";
+import type { BechAddr, BechAddr20, Option } from "lib/types";
+import { parseWithError } from "lib/utils";
+
+import { zCodeData, zCodesResponse } from "./types";
+import type { CodeData, CodesResponse } from "./types";
+
+const getCodes = async (
+  endpoint: string,
+  limit: number,
+  offset: number,
+  address: Option<BechAddr20>,
+  permission: Option<boolean>
+): Promise<CodesResponse> =>
+  axios
+    .get(`${endpoint}`, {
+      params: {
+        limit,
+        offset,
+        address,
+        permission,
+      },
+    })
+    .then(({ data }) => parseWithError(zCodesResponse, data));
+
+export const useCodes = (
+  limit: number,
+  offset: number,
+  address: Option<BechAddr20>,
+  permission: Option<boolean>,
+  options?: Pick<UseQueryOptions<CodesResponse>, "onSuccess">
+): UseQueryResult<CodesResponse> => {
+  const endpoint = useBaseApiRoute("codes");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.CODES, endpoint, limit, offset, address, permission],
+    async () => getCodes(endpoint, limit, offset, address, permission),
+    { retry: 1, refetchOnWindowFocus: false, ...options }
+  );
+};
+
+const getCodesByAddress = async (
+  endpoint: string,
+  address: BechAddr,
+  limit: number,
+  offset: number
+): Promise<CodesResponse> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(address)}/wasm/codes`, {
+      params: {
+        limit,
+        offset,
+      },
+    })
+    .then(({ data }) => zCodesResponse.parse(data));
+
+export const useCodesByAddress = (
+  address: BechAddr,
+  limit: number,
+  offset: number
+): UseQueryResult<CodesResponse> => {
+  const endpoint = useBaseApiRoute("accounts");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.CODES_BY_ADDRESS, endpoint, address, limit, offset],
+    async () => getCodesByAddress(endpoint, address, limit, offset),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getCodeDataByCodeId = async (
+  endpoint: string,
+  codeId: number,
+  isGov: boolean
+): Promise<CodeData> =>
+  axios
+    .get(`${endpoint}/${codeId}/info`, {
+      params: {
+        is_gov: isGov,
+      },
+    })
+    .then(({ data }) => parseWithError(zCodeData, data));
+
+export const useCodeDataByCodeId = (codeId: number, enabled = true) => {
+  const { enabled: isGov } = useGovConfig({ shouldRedirect: false });
+  const endpoint = useBaseApiRoute("codes");
+
+  return useQuery<CodeData>(
+    [CELATONE_QUERY_KEYS.CODE_DATA, endpoint, codeId, isGov],
+    async () => getCodeDataByCodeId(endpoint, codeId, isGov),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+      enabled,
+    }
+  );
+};