+++ src/lib/services/wasm/proposal/api.ts
@@ -1,0 +1,473 @@
+import type { UseQueryOptions, UseQueryResult } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
+import axios from "axios";
+import { useCallback } from "react";
+
+import { CELATONE_QUERY_KEYS, useBaseApiRoute } from "lib/app-provider";
+import { useAssetInfos } from "lib/services/assetService";
+import { useMovePoolInfos } from "lib/services/move/poolService";
+import { big, zProposalType } from "lib/types";
+import type {
+  BechAddr,
+  BechAddr20,
+  BechAddr32,
+  Coin,
+  Option,
+  ProposalParams,
+  ProposalStatus,
+  ProposalType,
+  ProposalValidatorVote,
+  ProposalVotesInfo,
+  ProposalVoteType,
+  Token,
+} from "lib/types";
+import {
+  coinToTokenWithValue,
+  deexponentify,
+  formatTokenWithValue,
+  getTokenLabel,
+  parseWithError,
+  snakeToCamel,
+} from "lib/utils";
+
+import {
+  zProposalAnswerCountsResponse,
+  zProposalDataResponse,
+  zProposalParamsResponse,
+  zProposalsResponse,
+  zProposalVotesInfoResponse,
+  zProposalVotesResponse,
+  zRelatedProposalsResponse,
+} from "./types";
+import type {
+  DepositParams,
+  DepositParamsInternal,
+  GovParams,
+  ProposalAnswerCountsResponse,
+  ProposalDataResponse,
+  ProposalsResponse,
+  ProposalValidatorVotesResponse,
+  ProposalVotesResponse,
+  RelatedProposalsResponse,
+  UploadAccess,
+  VotingParams,
+  VotingParamsInternal,
+} from "./types";
+
+const fetchGovDepositParams = (
+  lcdEndpoint: string
+): Promise<DepositParamsInternal> =>
+  axios
+    .get<{
+      deposit_params: DepositParams;
+    }>(`${lcdEndpoint}/cosmos/gov/v1beta1/params/deposit`)
+    .then(({ data }) => snakeToCamel(data.deposit_params));
+
+const fetchGovVotingParams = (
+  lcdEndpoint: string
+): Promise<VotingParamsInternal> =>
+  axios
+    .get<{
+      voting_params: VotingParams;
+    }>(`${lcdEndpoint}/cosmos/gov/v1beta1/params/voting`)
+    .then(({ data }) => snakeToCamel(data.voting_params));
+
+const fetchGovUploadAccessParams = async (
+  lcdEndpoint: string
+): Promise<UploadAccess> =>
+  axios.get(`${lcdEndpoint}/upload_access`).then(({ data }) => data);
+
+export const useGovParams = (): UseQueryResult<GovParams> => {
+  const lcdEndpoint = useBaseApiRoute("rest");
+  const cosmwasmEndpoint = useBaseApiRoute("cosmwasm");
+  const { data: assetInfos } = useAssetInfos({ withPrices: false });
+  const { data: movePoolInfos } = useMovePoolInfos({ withPrices: false });
+
+  const queryFn = useCallback(
+    () =>
+      Promise.all([
+        fetchGovDepositParams(lcdEndpoint),
+        fetchGovUploadAccessParams(cosmwasmEndpoint),
+        fetchGovVotingParams(lcdEndpoint),
+      ]).then<GovParams>((params) => {
+        const minDepositParam = params[0].minDeposit[0];
+        const minDepositToken = coinToTokenWithValue(
+          minDepositParam.denom,
+          minDepositParam.amount,
+          assetInfos,
+          movePoolInfos
+        );
+        const minDepositAmount = deexponentify(
+          minDepositToken.amount,
+          minDepositToken.precision
+        ).toFixed() as Token;
+
+        return {
+          depositParams: {
+            ...params[0],
+            minDeposit: {
+              ...minDepositParam,
+              amount: minDepositToken.amount,
+              formattedAmount: minDepositAmount,
+              formattedDenom: getTokenLabel(
+                minDepositToken.denom,
+                minDepositToken.symbol
+              ),
+              formattedToken: formatTokenWithValue(minDepositToken),
+              precision: minDepositToken.precision ?? 0,
+            },
+            minInitialDeposit: big(params[0].minInitialDepositRatio)
+              .times(minDepositAmount)
+              .toFixed(2) as Token,
+          },
+          uploadAccess: params[1],
+          votingParams: params[2],
+        };
+      }),
+    [assetInfos, cosmwasmEndpoint, lcdEndpoint, movePoolInfos]
+  );
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.GOV_PARAMS, lcdEndpoint, assetInfos],
+    queryFn,
+    {
+      keepPreviousData: true,
+      refetchOnWindowFocus: false,
+    }
+  );
+};
+
+export const useUploadAccessParams = (): UseQueryResult<UploadAccess> => {
+  const cosmwasmEndpoint = useBaseApiRoute("cosmwasm");
+  return useQuery(
+    [CELATONE_QUERY_KEYS.UPLOAD_ACCESS_PARAMS, cosmwasmEndpoint],
+    () => fetchGovUploadAccessParams(cosmwasmEndpoint),
+    { keepPreviousData: true, refetchOnWindowFocus: false }
+  );
+};
+
+const getProposalParams = async (
+  endpoint: string
+): Promise<ProposalParams<Coin>> =>
+  axios
+    .get(`${endpoint}/params`)
+    .then(({ data }) => parseWithError(zProposalParamsResponse, data));
+
+export const useProposalParams = () => {
+  const endpoint = useBaseApiRoute("proposals");
+  return useQuery<ProposalParams<Coin>>(
+    [CELATONE_QUERY_KEYS.PROPOSAL_PARAMS, endpoint],
+    async () => getProposalParams(endpoint),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getProposalTypes = async (endpoint: string) =>
+  axios
+    .get(`${endpoint}/types`)
+    .then(({ data }) => parseWithError(zProposalType.array(), data));
+
+export const useProposalTypes = () => {
+  const endpoint = useBaseApiRoute("proposals");
+  return useQuery<ProposalType[]>(
+    [CELATONE_QUERY_KEYS.PROPOSAL_TYPES, endpoint],
+    async () => getProposalTypes(endpoint),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getProposals = async (
+  endpoint: string,
+  limit: number,
+  offset: number,
+  proposer: Option<BechAddr20>,
+  statuses: ProposalStatus[],
+  types: ProposalType[],
+  search: string
+): Promise<ProposalsResponse> =>
+  axios
+    .get(`${endpoint}`, {
+      params: {
+        limit,
+        offset,
+        proposer,
+        statuses: statuses.join(","),
+        types: types.join(","),
+        search,
+      },
+    })
+    .then(({ data }) => parseWithError(zProposalsResponse, data));
+
+export const useProposals = (
+  limit: number,
+  offset: number,
+  proposer: Option<BechAddr20>,
+  statuses: ProposalStatus[],
+  types: ProposalType[],
+  search: string
+) => {
+  const endpoint = useBaseApiRoute("proposals");
+  const trimmedSearch = search.trim();
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS,
+      endpoint,
+      limit,
+      offset,
+      proposer,
+      statuses,
+      types,
+      trimmedSearch,
+    ],
+    async () =>
+      getProposals(
+        endpoint,
+        limit,
+        offset,
+        proposer,
+        statuses,
+        types,
+        trimmedSearch
+      ),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getProposalsByAddress = async (
+  endpoint: string,
+  address: BechAddr,
+  limit: number,
+  offset: number
+): Promise<ProposalsResponse> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(address)}/proposals`, {
+      params: {
+        limit,
+        offset,
+      },
+    })
+    .then(({ data }) => parseWithError(zProposalsResponse, data));
+
+export const useProposalsByAddress = (
+  address: BechAddr,
+  offset: number,
+  limit: number
+): UseQueryResult<ProposalsResponse> => {
+  const endpoint = useBaseApiRoute("accounts");
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS_BY_ADDRESS,
+      endpoint,
+      address,
+      limit,
+      offset,
+    ],
+    async () => getProposalsByAddress(endpoint, address, limit, offset),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getRelatedProposalsByContractAddress = async (
+  endpoint: string,
+  contractAddress: BechAddr32,
+  limit: number,
+  offset: number
+): Promise<RelatedProposalsResponse> =>
+  axios
+    .get(
+      `${endpoint}/${encodeURIComponent(contractAddress)}/related-proposals`,
+      {
+        params: {
+          limit,
+          offset,
+        },
+      }
+    )
+    .then(({ data }) => parseWithError(zRelatedProposalsResponse, data));
+
+export const useRelatedProposalsByContractAddress = (
+  contractAddress: BechAddr32,
+  offset: number,
+  limit: number
+) => {
+  const endpoint = useBaseApiRoute("contracts");
+
+  return useQuery<RelatedProposalsResponse>(
+    [
+      CELATONE_QUERY_KEYS.RELATED_PROPOSALS_BY_CONTRACT_ADDRESS,
+      endpoint,
+      contractAddress,
+      limit,
+      offset,
+    ],
+    async () =>
+      getRelatedProposalsByContractAddress(
+        endpoint,
+        contractAddress,
+        limit,
+        offset
+      ),
+    {
+      retry: 1,
+      keepPreviousData: true,
+    }
+  );
+};
+
+const getProposalData = async (
+  endpoint: string,
+  id: number
+): Promise<ProposalDataResponse> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(id)}/info`)
+    .then(({ data }) => parseWithError(zProposalDataResponse, data));
+
+export const useProposalData = (id: number, enabled = true) => {
+  const endpoint = useBaseApiRoute("proposals");
+
+  return useQuery<ProposalDataResponse>(
+    [CELATONE_QUERY_KEYS.PROPOSAL_DATA, endpoint, id],
+    async () => getProposalData(endpoint, id),
+    { retry: 1, enabled }
+  );
+};
+
+const getProposalVotesInfo = async (
+  endpoint: string,
+  id: number
+): Promise<ProposalVotesInfo> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(id)}/votes-info`)
+    .then(({ data }) => parseWithError(zProposalVotesInfoResponse, data));
+
+export const useProposalVotesInfo = (id: number) => {
+  const endpoint = useBaseApiRoute("proposals");
+
+  return useQuery<ProposalVotesInfo>(
+    [CELATONE_QUERY_KEYS.PROPOSAL_VOTES_INFO, endpoint, id],
+    async () => getProposalVotesInfo(endpoint, id),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+const getProposalVotes = async (
+  endpoint: string,
+  id: number,
+  limit: number,
+  offset: number,
+  answer: ProposalVoteType,
+  search: string
+): Promise<ProposalVotesResponse> => {
+  let url = `${endpoint}/${encodeURIComponent(id)}/votes?limit=${limit}&offset=${offset}`;
+  url = url.concat(search ? `&search=${encodeURIComponent(search)}` : "");
+  url = url.concat(answer ? `&answer=${encodeURIComponent(answer)}` : "");
+
+  return axios
+    .get(url)
+    .then(({ data }) => parseWithError(zProposalVotesResponse, data));
+};
+
+export const useProposalVotes = (
+  id: number,
+  limit: number,
+  offset: number,
+  answer: ProposalVoteType,
+  search: string,
+  options: Pick<UseQueryOptions<ProposalVotesResponse>, "onSuccess"> = {}
+): UseQueryResult<ProposalVotesResponse> => {
+  const endpoint = useBaseApiRoute("proposals");
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSAL_VOTES,
+      endpoint,
+      id,
+      limit,
+      offset,
+      search,
+      answer,
+    ],
+    async () => getProposalVotes(endpoint, id, limit, offset, answer, search),
+    { retry: 1, refetchOnWindowFocus: false, ...options }
+  );
+};
+
+const getProposalValidatorVotes = async (
+  endpoint: string,
+  id: number,
+  limit: number,
+  offset: number,
+  answer: ProposalVoteType,
+  search: string
+): Promise<ProposalValidatorVotesResponse> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(id)}/validator-votes`, {
+      params: {
+        limit,
+        offset,
+        answer,
+        search,
+      },
+    })
+    .then(({ data }) => {
+      const parsed = parseWithError(zProposalVotesResponse, data);
+      return {
+        items: parsed.items.map<ProposalValidatorVote>((item, idx) => ({
+          ...item,
+          rank: idx + 1,
+        })),
+        total: parsed.total,
+      };
+    });
+
+export const useProposalValidatorVotes = (
+  id: number,
+  limit: number,
+  offset: number,
+  answer: ProposalVoteType,
+  search: string,
+  options: Pick<
+    UseQueryOptions<ProposalValidatorVotesResponse>,
+    "onSuccess"
+  > = {}
+) => {
+  const endpoint = useBaseApiRoute("proposals");
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSAL_VALIDATOR_VOTES,
+      endpoint,
+      id,
+      limit,
+      offset,
+      answer,
+      search,
+    ],
+    async () =>
+      getProposalValidatorVotes(endpoint, id, limit, offset, answer, search),
+    { retry: 1, refetchOnWindowFocus: false, ...options }
+  );
+};
+
+const getProposalAnswerCounts = async (
+  endpoint: string,
+  id: number
+): Promise<ProposalAnswerCountsResponse> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(id)}/answer-counts`)
+    .then(({ data }) => parseWithError(zProposalAnswerCountsResponse, data));
+
+export const useProposalAnswerCounts = (
+  id: number,
+  validatorOnly = false
+): UseQueryResult<ProposalAnswerCountsResponse> => {
+  const endpoint = useBaseApiRoute("proposals");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.PROPOSAL_ANSWER_COUNTS, endpoint, id, validatorOnly],
+    async () => getProposalAnswerCounts(endpoint, id),
+    { retry: 1, refetchOnWindowFocus: false }
+  );
+};