+++ src/lib/utils/proposal.ts
@@ -1,6 +1,6 @@
 import type { Coin } from "@cosmjs/stargate";
 
-import type { MinDeposit } from "lib/services/proposalService";
+import type { MinDeposit } from "lib/services/wasm/proposal";
 import { big } from "lib/types";
 import type { Option } from "lib/types";